Refactor Mongo compator to remove SingleBucketStorage and related abstractions.
