+++ modules/module-mongodb-storage/src/storage/implementation/v1/SingleBucketStoreV1.ts
@@ -1,58 +1,50 @@
 import { mongo } from '@powersync/lib-service-mongodb';
 import { InternalOpId } from '@powersync/service-core';
 import { BucketDataDoc, BucketKey } from '../common/BucketDataDoc.js';
-import {
-  BucketDataDocumentGeneric,
-  BucketDataDocumentGenericId,
-  SingleBucketStore
-} from '../common/SingleBucketStore.js';
 import { BucketDataProperties } from '../models.js';
 import { VersionedPowerSyncMongoV1 } from './VersionedPowerSyncMongoV1.js';
 import { BucketDataDocumentV1, BucketDataKeyV1, serializeBucketDataV1 } from './models.js';
 
-export class SingleBucketStoreV1 implements SingleBucketStore {
-  public readonly collection: mongo.Collection<BucketDataDocumentGeneric>;
+export class SingleBucketStoreV1 {
+  public readonly collection: mongo.Collection<BucketDataDocumentV1>;
 
   constructor(
-    private db: VersionedPowerSyncMongoV1,
+    db: VersionedPowerSyncMongoV1,
     public readonly key: BucketKey
   ) {
-    this.collection = db.bucketDataV1 as unknown as mongo.Collection<BucketDataDocumentGeneric>;
+    this.collection = db.bucketDataV1;
   }
 
-  docId(o: InternalOpId): BucketDataDocumentGenericId {
-    // `satisfies BucketDataKeyV1` checks that we use the correct type for V1 storage
-    // `as anyt` is to allow casting to the interface virtual type
+  docId(o: InternalOpId): BucketDataKeyV1 {
     return {
       g: this.key.replicationStreamId,
       b: this.key.bucket,
       o
-    } satisfies BucketDataKeyV1 as any;
+    };
   }
 
-  get minId(): BucketDataDocumentGenericId {
+  get minId(): BucketDataKeyV1 {
     return {
       g: this.key.replicationStreamId,
       b: this.key.bucket,
-      o: new mongo.MinKey()
-    } as any;
+      o: new mongo.MinKey() as any
+    };
   }
 
-  get maxId(): BucketDataDocumentGenericId {
+  get maxId(): BucketDataKeyV1 {
     return {
       g: this.key.replicationStreamId,
       b: this.key.bucket,
-      o: new mongo.MaxKey()
-    } as any;
+      o: new mongo.MaxKey() as any
+    };
   }
 
-  toPersistedDocument(source: Omit<BucketDataDoc, 'bucketKey'>): BucketDataDocumentGeneric {
-    return serializeBucketDataV1({ bucketKey: this.key, ...source }) as unknown as BucketDataDocumentGeneric;
+  toPersistedDocument(source: Omit<BucketDataDoc, 'bucketKey'>): BucketDataDocumentV1 {
+    return serializeBucketDataV1({ bucketKey: this.key, ...source });
   }
 
-  fromPersistedDocument(doc: BucketDataDocumentGeneric): BucketDataDoc {
-    const document = doc as unknown as BucketDataDocumentV1;
-    const { _id, ...rest } = document;
+  fromPersistedDocument(doc: BucketDataDocumentV1): BucketDataDoc {
+    const { _id, ...rest } = doc;
     return {
       bucketKey: this.key,
       o: _id.o,
@@ -61,10 +53,9 @@ export class SingleBucketStoreV1 implements SingleBucketStore {
   }
 
   fromPartialPersistedDocument<T extends keyof BucketDataProperties>(
-    doc: Pick<BucketDataDocumentGeneric, '_id' | T>
+    doc: Pick<BucketDataDocumentV1, '_id' | T>
   ): Pick<BucketDataDoc, 'bucketKey' | 'o' | T> {
-    const document = doc as Pick<BucketDataDocumentV1, '_id' | T>;
-    const { _id, ...rest } = document;
+    const { _id, ...rest } = doc;
     return {
       bucketKey: this.key,
       o: _id.o,
